examples/terminal-agent/pyproject.toml
[project]
name = "terminal-agent"
version = "0.1.0"
description = "Claude Code-esque terminal agent for infrastructure automation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2",
    "typing-extensions>=4.7.1",
    "python-dotenv>=1.0.0",
    "textual>=0.50.0",
    "virsh-sandbox>=0.0.21b0",
    "pyyaml>=6.0",
    "ansible>=13.2.0",
    "ansible-lint>=26.1.0",
    "mcp>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio==0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
strict = true
